requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.7.2
asyncio==3.4.3
attrs==23.2.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.1
defusedxml==0.7.1
dhanhq==1.2.3
distlib==0.3.8
Django==5.0.6
Django==3.2.20           
djangorestframework==3.12.4  # or your desired version
psycopg2-binary==2.9.3   
drf-yasg==1.20.0         
python-decouple==3.7     
django-filter==21.1       
pytz==2021.3  
factory_boy==3.3.0
pytest
pytest-django